+++ src/ladb_opencutlist/ruby/constants.rb
@@ -2,6 +2,6 @@ module Ladb::OpenCutList
 
     EXTENSION_NAME = 'OpenCutList'.freeze
     EXTENSION_VERSION = '1.9.5-dev'.freeze
-    EXTENSION_BUILD = '202009271847'.freeze
+    EXTENSION_BUILD = '202009281001'.freeze
 
 end